requirements.txt
kivy
pycryptodome
